# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebDev Orders API",
    "version": "1.0.0",
    "description": "API документация для системы управления заявками на веб-разработку"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.webdev-orders.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Application": {
        "type": "object",
        "required": [
          "title",
          "service_type",
          "contact_email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Уникальный идентификатор заявки",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID пользователя, создавшего заявку",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "title": {
            "type": "string",
            "description": "Название заявки",
            "example": "Создание корпоративного сайта"
          },
          "description": {
            "type": "string",
            "description": "Описание заявки",
            "example": "Необходимо создать современный корпоративный сайт с CMS"
          },
          "service_type": {
            "type": "string",
            "enum": [
              "landing_page",
              "corporate_site",
              "ecommerce",
              "web_application",
              "other"
            ],
            "description": "Тип услуги",
            "example": "corporate_site"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "description": "Контактный email",
            "example": "client@example.com"
          },
          "contact_phone": {
            "type": "string",
            "description": "Контактный телефон",
            "example": 79991234567
          },
          "company_name": {
            "type": "string",
            "description": "Название компании клиента",
            "example": "ООО \"Рога и копыта\""
          },
          "expected_budget": {
            "type": "integer",
            "description": "Ожидаемый бюджет в рублях",
            "example": 150000
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "in_review",
              "approved",
              "in_progress",
              "completed",
              "cancelled",
              "rejected"
            ],
            "description": "Статус заявки",
            "example": "draft"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "normal",
              "high",
              "urgent"
            ],
            "description": "Приоритет заявки",
            "example": "normal"
          },
          "assigned_to": {
            "type": "string",
            "format": "uuid",
            "description": "ID менеджера, назначенного на заявку",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата подачи заявки",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "description": "Крайний срок выполнения",
            "example": "2023-02-01T00:00:00.000Z"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата создания заявки",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата последнего обновления",
            "example": "2023-01-01T00:00:00.000Z"
          }
        }
      },
...
    }
  },
  "paths": {
    "/admin/applications": {
      "get": {
        "summary": "Получение списка заявок с фильтрацией",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по статусу заявки"
          },
          {
            "in": "query",
            "name": "service_type",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по типу услуги"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по приоритету"
          },
          {
            "in": "query",
            "name": "assigned_to",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по назначенным менеджерам"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Номер страницы"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Количество элементов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Список заявок",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "applications": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Application"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/applications/{id}": {
      "get": {
        "summary": "Получение детальной информации о заявке",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID заявки"
          }
        ],
        "responses": {
          "200": {
            "description": "Детальная информация о заявке",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "application": {
                          "$ref": "#/components/schemas/Application"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заявка не найдена"
          }
        }
      }
    },
    "/admin/applications/{id}/status": {
      "put": {
        "summary": "Изменение статуса заявки",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID заявки"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "submitted",
                      "in_review",
                      "approved",
                      "in_progress",
                      "completed",
                      "cancelled",
                      "rejected"
                    ],
                    "description": "Новый статус заявки"
                  },
                  "comment": {
                    "type": "string",
                    "description": "Комментарий к изменению статуса"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Статус заявки успешно изменен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "application": {
                          "$ref": "#/components/schemas/Application"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "404": {
            "description": "Заявка не найдена"
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "summary": "Получение списка пользователей",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по роли пользователя"
          },
          {
            "in": "query",
            "name": "is_active",
            "schema": {
              "type": "boolean"
            },
            "description": "Фильтр по активности пользователя"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Номер страницы"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Количество элементов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Список пользователей",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/applications/{id}/reset-to-draft": {
      "post": {
        "summary": "Возврат заявки в черновик",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID заявки"
          }
        ],
        "responses": {
          "200": {
            "description": "Заявка успешно возвращена в черновик",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "application": {
                          "$ref": "#/components/schemas/Application"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "404": {
            "description": "Заявка не найдена"
          }
        }
      }
    },
    "/admin/applications/{id}/assign": {
      "put": {
        "summary": "Назначение менеджера на заявку",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID заявки"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "manager_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID менеджера",
                    "example": "123e4567-e89b-12d3-a456-426614174002"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Менеджер успешно назначен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "application_id": {
                          "type": "string"
                        },
                        "old_manager_id": {
                          "type": "string"
                        },
                        "new_manager": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "404": {
            "description": "Заявка или менеджер не найдены"
          }
        }
      }
    },
    "/admin/applications/{id}/notes": {
      "post": {
        "summary": "Добавление внутренней заметки к заявке",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID заявки"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": {
                    "type": "string",
                    "description": "Содержание заметки",
                    "example": "Важное замечание по заявке"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Заметка успешно добавлена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "note": {
                          "$ref": "#/components/schemas/Note"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          }
        }
      }
    },
    "/admin/users/{id}/role": {
      "put": {
        "summary": "Изменение роли пользователя",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID пользователя"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "client",
                      "manager",
                      "admin",
                      "super_admin"
                    ],
                    "description": "Новая роль пользователя",
                    "example": "manager"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Роль пользователя успешно изменена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "full_name": {
                              "type": "string"
                            },
                            "old_role": {
                              "type": "string"
                            },
                            "new_role": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        }
      }
    },
    "/admin/stats/dashboard": {
      "get": {
        "summary": "Получение статистики для дашборда",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Статистика для дашборда",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "stats": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}